+++ fabric/src/main/java/mod/fuji/module/initializer/rtp/RtpInitializer.java
@@ -16,6 +16,7 @@ import mod.fuji.core.config.handler.impl.ObjectConfigurationHandler;
 import mod.fuji.core.service.random_teleport.RandomTeleporter;
 import mod.fuji.core.service.random_teleport.structure.RandomTeleportSettings;
 import mod.fuji.module.initializer.ModuleInitializer;
+import mod.fuji.module.initializer.rtp.command.argument.wrapper.RtpDimension;
 import mod.fuji.module.initializer.rtp.config.model.RtpConfigModel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.server.level.ServerLevel;
@@ -59,7 +60,7 @@ public class RtpInitializer extends ModuleInitializer {
 
     private static final BaseConfigurationHandler<RtpConfigModel> config = ObjectConfigurationHandler.ofModule(BaseConfigurationHandler.CONFIG_JSON_LITERAL, RtpConfigModel.class);
 
-    private static Optional<RandomTeleportSettings> getRandomTeleportSettings(@NotNull ServerLevel world) {
+    public static Optional<RandomTeleportSettings> getRandomTeleportSettings(@NotNull ServerLevel world) {
         List<RandomTeleportSettings> list = config.model().getDimensions().getSettings();
         String dimension = RegistryHelper.getIdAsString(world);
         return list.stream()
@@ -70,7 +71,7 @@ public class RtpInitializer extends ModuleInitializer {
 
     @Document(id = 1751826340406L, value = "Random rtp in specified dimension.")
     @CommandNode("rtp")
-    private static int $rtp(@CommandSource @CommandTarget ServerPlayer player, Optional<Dimension> dimension) {
+    private static int $rtp(@CommandSource @CommandTarget ServerPlayer player, Optional<RtpDimension> dimension) {
         ServerLevel serverWorld = dimension
             .map(Dimension::getValue)
             .orElseGet(() -> EntityHelper.getServerWorld(player));