+++ fabric/src/main/java/mod/fuji/module/initializer/rtp/command/argument/wrapper/RtpDimension.java
@@ -1,0 +1,10 @@
+package mod.fuji.module.initializer.rtp.command.argument.wrapper;
+
+import mod.fuji.core.command.argument.wrapper.impl.Dimension;
+import net.minecraft.server.level.ServerLevel;
+
+public class RtpDimension extends Dimension {
+    public RtpDimension(ServerLevel value) {
+        super(value);
+    }
+}