+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ## ✨ Features
 - [command_bundle] feature: add new command definitions (`/success` and `/failure`) into the `default config` model.
+- [rtp] feature: improve the command suggestion for `--dimension` optional argument. Now it only suggests the effective dimensions.
 
 ## 🐞 Fixes
 - [chat.style] When installed on the client-side, re-entering the single-player world would fail.